feat(command): add an elevator goto command
This should be a better solution instead of just moving up and down as soon as we hit the button.

Instead, we set the desired stage and then we actually move to it. This should avoid queue issues along with other random stuff like that.

GotoElevatorCommand is basically the same as CorrectElevatorCommand, so I might just get rid of that. I like the sag solution though, so I might base GotoElevatorCommand off stageDrift instead of just comparing desiredStage and currentStage, and all I need to do is copy CorrectElevatorCommand but add support for going down.